# Metric refactor - 2x perf and allocation free