--- a/opentelemetry-sdk/src/metrics/internal/sum.rs
+++ b/opentelemetry-sdk/src/metrics/internal/sum.rs
@@ -1,4 +1,6 @@
-use std::sync::atomic::{AtomicBool, Ordering};
+use std::collections::HashSet;
+use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
+use std::sync::Arc;
 use std::vec;
 use std::{
     collections::HashMap,
@@ -8,19 +10,21 @@ use std::{
 
 use crate::metrics::data::{self, Aggregation, DataPoint, Temporality};
 use crate::metrics::AttributeSet;
+use once_cell::sync::Lazy;
 use opentelemetry::KeyValue;
 use opentelemetry::{global, metrics::MetricsError};
 
-use super::{
-    aggregate::{is_under_cardinality_limit, STREAM_OVERFLOW_ATTRIBUTE_SET},
-    AtomicTracker, Number,
-};
+use super::{aggregate::is_under_cardinality_limit, AtomicTracker, Number};
+
+pub(crate) static STREAM_OVERFLOW_ATTRIBUTES: Lazy<Vec<KeyValue>> =
+    Lazy::new(|| vec![KeyValue::new("otel.metric.overflow", "true")]);
 
 /// The storage for sums.
 struct ValueMap<T: Number<T>> {
-    values: RwLock<HashMap<AttributeSet, T::AtomicTracker>>,
+    values: RwLock<HashMap<Vec<KeyValue>, Arc<T::AtomicTracker>>>,
     has_no_value_attribute_value: AtomicBool,
     no_attribute_value: T::AtomicTracker,
+    count: AtomicUsize,
 }
 
 impl<T: Number<T>> Default for ValueMap<T> {
@@ -35,42 +39,65 @@ impl<T: Number<T>> ValueMap<T> {
             values: RwLock::new(HashMap::new()),
             has_no_value_attribute_value: AtomicBool::new(false),
             no_attribute_value: T::new_atomic_tracker(),
+            count: AtomicUsize::new(0),
         }
     }
 }
 
 impl<T: Number<T>> ValueMap<T> {
-    fn measure(&self, measurement: T, attrs: AttributeSet) {
+    fn measure(&self, measurement: T, attrs: &[KeyValue]) {
         if attrs.is_empty() {
             self.no_attribute_value.add(measurement);
             self.has_no_value_attribute_value
                 .store(true, Ordering::Release);
         } else if let Ok(values) = self.values.read() {
-            if let Some(value_to_update) = values.get(&attrs) {
+            // Try incoming order first
+            if let Some(value_to_update) = values.get(attrs) {
                 value_to_update.add(measurement);
                 return;
             } else {
-                drop(values);
-                if let Ok(mut values) = self.values.write() {
-                    // Recheck after acquiring write lock, in case another
-                    // thread has added the value.
-                    if let Some(value_to_update) = values.get(&attrs) {
-                        value_to_update.add(measurement);
-                        return;
-                    } else if is_under_cardinality_limit(values.len()) {
-                        let new_value = T::new_atomic_tracker();
-                        new_value.add(measurement);
-                        values.insert(attrs, new_value);
-                    } else if let Some(overflow_value) =
-                        values.get_mut(&STREAM_OVERFLOW_ATTRIBUTE_SET)
-                    {
-                        overflow_value.add(measurement);
-                        return;
-                    } else {
-                        let new_value = T::new_atomic_tracker();
-                        new_value.add(measurement);
-                        values.insert(STREAM_OVERFLOW_ATTRIBUTE_SET.clone(), new_value);
-                        global::handle_error(MetricsError::Other("Warning: Maximum data points for metric stream exceeded. Entry added to overflow. Subsequent overflows to same metric until next collect will not be logged.".into()));
+                // Then try sorted order.                
+                let sorted_attrs = AttributeSet::from(attrs).into_vec();
+                if let Some(value_to_update) = values.get(sorted_attrs.as_slice()) {
+                    value_to_update.add(measurement);
+                    return;
+                } else {
+                    // Give up the lock, before acquiring write lock.
+                    drop(values);
+                    if let Ok(mut values) = self.values.write() {
+                        // Recheck both incoming and sorted after acquiring
+                        // write lock, in case another thread has added the
+                        // value.
+                        if let Some(value_to_update) = values.get(attrs) {
+                            value_to_update.add(measurement);
+                            return;
+                        } else if let Some(value_to_update) = values.get(sorted_attrs.as_slice()) {
+                            value_to_update.add(measurement);
+                            return;
+                        } else if is_under_cardinality_limit(self.count.load(Ordering::Acquire)) {
+                            let new_value = T::new_atomic_tracker();
+                            new_value.add(measurement);
+                            let new_value = Arc::new(new_value);
+
+                            // Insert original order
+                            values.insert(attrs.to_vec(), new_value.clone());
+                            
+                            // Insert sorted order
+                            values.insert(sorted_attrs, new_value);
+
+                            self.count.fetch_add(1, Ordering::Release);
+                            
+                        } else if let Some(overflow_value) =
+                            values.get_mut(STREAM_OVERFLOW_ATTRIBUTES.as_slice())
+                        {                            
+                            overflow_value.add(measurement);
+                            return;
+                        } else {
+                            let new_value = T::new_atomic_tracker();
+                            new_value.add(measurement);
+                            values.insert(STREAM_OVERFLOW_ATTRIBUTES.clone(), Arc::new(new_value));
+                            global::handle_error(MetricsError::Other("Warning: Maximum data points for metric stream exceeded. Entry added to overflow. Subsequent overflows to same metric until next collect will not be logged.".into()));
+                        }
                     }
                 }
             }
@@ -100,7 +127,6 @@ impl<T: Number<T>> Sum<T> {
     }
 
     pub(crate) fn measure(&self, measurement: T, attrs: &[KeyValue]) {
-        let attrs: AttributeSet = attrs.into();
         self.value_map.measure(measurement, attrs)
     }
 
@@ -124,20 +150,17 @@ impl<T: Number<T>> Sum<T> {
         s_data.temporality = Temporality::Delta;
         s_data.is_monotonic = self.monotonic;
         s_data.data_points.clear();
-
-        let mut values = match self.value_map.values.write() {
-            Ok(v) => v,
-            Err(_) => return (0, None),
-        };
-
-        let n = values.len() + 1;
+        
+        // Max number of data points need to account for the special casing
+        // of the no attribute value + overflow attribute.
+        let n = self.value_map.count.load(Ordering::Relaxed) + 2;
         if n > s_data.data_points.capacity() {
             s_data
                 .data_points
                 .reserve_exact(n - s_data.data_points.capacity());
         }
 
-        let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);
+        let prev_start = self.start.lock().map(|start| *start).unwrap_or(t);        
         if self
             .value_map
             .has_no_value_attribute_value
@@ -152,23 +175,29 @@ impl<T: Number<T>> Sum<T> {
             });
         }
 
+        let mut values = match self.value_map.values.write() {
+            Ok(v) => v,
+            Err(_) => return (0, None),
+        };
+
+        let mut seen = HashSet::new();
         for (attrs, value) in values.drain() {
-            s_data.data_points.push(DataPoint {
-                attributes: attrs
-                    .iter()
-                    .map(|(k, v)| KeyValue::new(k.clone(), v.clone()))
-                    .collect(),
-                start_time: Some(prev_start),
-                time: Some(t),
-                value: value.get_value(),
-                exemplars: vec![],
-            });
+            if seen.insert(Arc::as_ptr(&value)) {
+                s_data.data_points.push(DataPoint {
+                    attributes: attrs.clone(),
+                    start_time: Some(prev_start),
+                    time: Some(t),
+                    value: value.get_value(),
+                    exemplars: vec![],
+                });
+            }
         }
 
         // The delta collection cycle resets.
         if let Ok(mut start) = self.start.lock() {
             *start = t;
         }
+        self.value_map.count.store(0, Ordering::Release);
 
         (
             s_data.data_points.len(),
@@ -231,10 +260,7 @@ impl<T: Number<T>> Sum<T> {
         // overload the system.
         for (attrs, value) in values.iter() {
             s_data.data_points.push(DataPoint {
-                attributes: attrs
-                    .iter()
-                    .map(|(k, v)| KeyValue::new(k.clone(), v.clone()))
-                    .collect(),
+                attributes: attrs.clone(),
                 start_time: Some(prev_start),
                 time: Some(t),
                 value: value.get_value(),
@@ -254,7 +280,7 @@ pub(crate) struct PrecomputedSum<T: Number<T>> {
     value_map: ValueMap<T>,
     monotonic: bool,
     start: Mutex<SystemTime>,
-    reported: Mutex<HashMap<AttributeSet, T>>,
+    reported: Mutex<HashMap<Vec<KeyValue>, T>>,
 }
 
 impl<T: Number<T>> PrecomputedSum<T> {
@@ -268,7 +294,6 @@ impl<T: Number<T>> PrecomputedSum<T> {
     }
 
     pub(crate) fn measure(&self, measurement: T, attrs: &[KeyValue]) {
-        let attrs: AttributeSet = attrs.into();
         self.value_map.measure(measurement, attrs)
     }
 
@@ -332,10 +357,7 @@ impl<T: Number<T>> PrecomputedSum<T> {
                 new_reported.insert(attrs.clone(), value.get_value());
             }
             s_data.data_points.push(DataPoint {
-                attributes: attrs
-                    .iter()
-                    .map(|(k, v)| KeyValue::new(k.clone(), v.clone()))
-                    .collect(),
+                attributes: attrs.clone(),
                 start_time: Some(prev_start),
                 time: Some(t),
                 value: delta,
@@ -347,6 +369,7 @@ impl<T: Number<T>> PrecomputedSum<T> {
         if let Ok(mut start) = self.start.lock() {
             *start = t;
         }
+        self.value_map.count.store(0, Ordering::Release);
 
         *reported = new_reported;
         drop(reported); // drop before values guard is dropped
@@ -417,10 +440,7 @@ impl<T: Number<T>> PrecomputedSum<T> {
                 new_reported.insert(attrs.clone(), value.get_value());
             }
             s_data.data_points.push(DataPoint {
-                attributes: attrs
-                    .iter()
-                    .map(|(k, v)| KeyValue::new(k.clone(), v.clone()))
-                    .collect(),
+                attributes: attrs.clone(),
                 start_time: Some(prev_start),
                 time: Some(t),
                 value: delta,