--- a/opentelemetry-sdk/src/metrics/internal/sum.rs
+++ b/opentelemetry-sdk/src/metrics/internal/sum.rs
@@ -218,7 +218,7 @@ impl<T: Number<T>> Sum<T> {
         let s_data = s_data.unwrap_or_else(|| new_agg.as_mut().expect("present if s_data is none"));
         s_data.temporality = Temporality::Cumulative;
         s_data.is_monotonic = self.monotonic;
-        s_data.data_points.clear();        
+        s_data.data_points.clear();
 
         // Max number of data points need to account for the special casing
         // of the no attribute value + overflow attribute.
@@ -313,7 +313,7 @@ impl<T: Number<T>> PrecomputedSum<T> {
         let s_data = s_data.unwrap_or_else(|| new_agg.as_mut().expect("present if s_data is none"));
         s_data.data_points.clear();
         s_data.temporality = Temporality::Delta;
-        s_data.is_monotonic = self.monotonic;        
+        s_data.is_monotonic = self.monotonic;
 
         // Max number of data points need to account for the special casing
         // of the no attribute value + overflow attribute.
@@ -398,7 +398,7 @@ impl<T: Number<T>> PrecomputedSum<T> {
         let s_data = s_data.unwrap_or_else(|| new_agg.as_mut().expect("present if s_data is none"));
         s_data.data_points.clear();
         s_data.temporality = Temporality::Cumulative;
-        s_data.is_monotonic = self.monotonic;        
+        s_data.is_monotonic = self.monotonic;
 
         // Max number of data points need to account for the special casing
         // of the no attribute value + overflow attribute.