… opentelemetry-sdk/src/metrics/mod.rs
… stress/src/metrics_counter.rs