Part 2 of https://github.com/open-telemetry/opentelemetry-rust/issues/1740, addressing 1,2,3. Also closes #1954 for SUM.

Refactoring SUM aggregation logic to operate entirely on user provided slice of KeyValues. In hot-path, we avoid **allocating, copying, sorting, de-duplicating.**, leading to **2X or more gains** (with 4 attributes. More attributes = much more gains), as long as user do not keep changing the order of attributes.

This is mostly based on the prototypes did outside this repo in https://github.com/cijothomas/metrics-mini/tree/main/metrics
Once the approach is settled, we can replicate to all aggregations. (This PR is only fixing SUM)

Benchmarks:
```
Counter_Add_Sorted      time:   [192.64 ns 193.03 ns 193.48 ns]
                        change: [-66.328% -66.034% -65.769%] (p = 0.00 < 0.05)
                        Performance has improved.
Counter_Add_Unsorted    time:   [209.09 ns 209.41 ns 209.74 ns]
                        change: [-63.198% -62.905% -62.670%] (p = 0.00 < 0.05)
                        Performance has improved.
Counter_Overflow        time:   [895.61 ns 898.58 ns 902.11 ns]
                        change: [+54.000% +58.218% +61.362%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

Stress Test also shows small gains (17M/sec to 20M/sec). https://github.com/open-telemetry/opentelemetry-rust/pull/1833 fixed the major contention, so this PR has limited impact on stress test, but the raw speed is significantly improved.

**Stress Test with memory stats:**

Main:
Throughput: 17,842,800 iterations/sec
Memory usage: 5.06 MB
CPU usage: 98.61875%
Virtual memory usage: 2249.73 MB

PR:
Throughput: 18,869,600 iterations/sec
Memory usage: 4.23 MB
CPU usage: 99.2016%
Virtual memory usage: 2249.55 MB

Visible memory reduction as well from ~5 MB to ~4.2 MB

*note*: Overflow perf regressed, and that is acceptable. The main goal with overflow feature is to protect SDK from consuming unlimited memory, so it is acceptable that this path is slower. This is rare case, and indicates user needs to adjust cardinality.